+++ src/imio/events/core/tests/test_entity.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from eea.facetednavigation.subtypes.interfaces import IFacetedNavigable
 from imio.events.core.contents.entity.content import IEntity  # NOQA E501
 from imio.events.core.testing import IMIO_EVENTS_CORE_INTEGRATION_TESTING  # noqa
 from plone import api
@@ -69,6 +70,8 @@ class TestEntity(unittest.TestCase):
                 obj.id,
             ),
         )
+
+        self.assertTrue(IFacetedNavigable.providedBy(obj))
 
         parent = obj.__parent__
         self.assertIn("imio.events.Entity", parent.objectIds())